Ändert die Addieren-Funktion von int auf float

+++ Program.cs
@@ -10,18 +10,20 @@ namespace Taschenrechner_Iteration_1
     {
         static void Main(string[] args)
         {
-            //Als Benutzer möchte ich zwei Zahlen eingeben, um deren Summe berechnen zu lassen
+            // Als Benutzer möchte ich zwei Gleitkommazahlen eingeben, um deren Summe berechnen zu lassen
+            // Floatzahlen sind Zahlen mit Komma...sie sind abe nur bis auf 10 Stellen genau
+
             Console.Write("Bitte gib den ersten Summanden ein: ");
             string ersterSummand = Console.ReadLine();
             Console.Write("Bitte gib den zweiten Summanden ein: ");
             string zweiterSummand = Console.ReadLine();
 
-            //Wandel Text in Ganzzahlen
-            int ersterSummandalsZahl = Convert.ToInt32(ersterSummand);
-            int zweiterSummandalsZahl = Convert.ToInt32(zweiterSummand);
+            //Wandel Text in Gleitkommazahlen
+            float ersterSummandalsZahl = Convert.ToSingle(ersterSummand);
+            float zweiterSummandalsZahl = Convert.ToSingle(zweiterSummand);
 
             //Berechnung ausführen
-            int summe = ersterSummandalsZahl + zweiterSummandalsZahl;
+            float summe = ersterSummandalsZahl + zweiterSummandalsZahl;
 
             Console.WriteLine($"Die Summe beider Summanden ist: {summe}");
             Console.ReadLine();
